refactor(persistence): move log datasource to persistence module and add tests
Implement LogDatasource with Prisma integration and validation
Add comprehensive test suite for LogDatasource functionality
Update imports to reflect new module structure
