+++ src/infrastructure/test-infrastructure.ts
@@ -1,5 +1,5 @@
 import { CreateLogValidator } from "./validators/create-log.validator";
-import { LogDatasource } from "./datasources/log.datasource";
+import { LogDatasource } from "./persistence/datasource/log.datasource";
 import { DatabaseClient } from "@/infrastructure/persistence/database.client";
 import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
